+++ pm/CHANGELOG.md
@@ -9,6 +9,7 @@ description: A detailed, chronological record of all changes made to a project, 
 ### Fixed
 
 - Added rate limit for various endpoints
+- Fixed draft available status code
 - Fixed discover search
 - Fixed waving hand icon
 - Fixed missing send message button